refactor : 리뷰반영하여 class 분리
