+++ docs/피드백정리.md
@@ -18,13 +18,14 @@
 
 - [x] function paymentCost() 함수명을 동사로 나타내보자 
   - handlePayment는 어떨까?
-- [] index.html에서 24라인 : form태그로 감싸면 onSubmit이벤트 하나로 pay동작을 핸들링 가능하다.
+- [x] index.html에서 24라인 : form태그로 감싸면 onSubmit이벤트 하나로 pay동작을 핸들링 가능하다.
+- [x] index.js 12라인, 전역변수를 사용한다면 유지보수에 어려움이 있다. class나 함수를 통해 변수가 변경되는 범위를 제한해보자.
+- [] index.js 50라인 : keyDown이벤트를 통해 key를 감지
 - [x] index.js 101라인 : 어떤 토글을 클릭했는지, 어떤 동작을 제어했는지 네이밍에 나타내본다.
-- [] index.js 124~130라인 네이밍 수정
+- [] index.js 124~130라인 네이밍 수정 (prefix고려해보기)
 - [x] index.js 55라인 digitRange : 검증하는 내용을 함수명에 담아본다.
   - validatePaymentInput
 - [x] makeLottoByUser() : 로또를 몇개 구매했는지를 나타내는 함수로, 함수명을 바꿔보자 - purchaseTicketNumber
-- [] index.js 12라인, 전역변수를 사용한다면 유지보수에 어려움이 있다. class나 함수를 통해 변수가 변경되는 범위를 제한해보자.
 - [x] index.js 94 ~ 100라인 : Array 메서드를 활용하면 간결하게 코드를 구성할 수 있다.
 - [x] index.js 104 ~ 116라인 : for문보다 forEach, map등 Array메서드를 사용하자
 - [x] lotto.spec.js 2~4라인 : beforeEach를 사용한다면 테스트마다 visit이 수행되도록 할 수 있다.
